# docs/openapi/raven-services.yaml
openapi: "3.1.0"
info:
  title: Raven AI — Microservice API
  version: "1.0.0"
  description: |
    Enterprise-grade personal AI assistant.
    Monolith API at :18888, microservices behind gateway at :8000.
servers:
  - url: http://localhost:18888/api
    description: Monolith API
  - url: http://localhost:8000/api/v1
    description: Gateway (microservice mode)
paths:
...
  /monitors:
    get:
      summary: List monitors
      tags: [Monitor]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Monitor list
    post:
      summary: Create monitor
      tags: [Monitor]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                url: { type: string, format: uri }
                interval_seconds: { type: integer, minimum: 10 }
                timeout_seconds: { type: integer, minimum: 1 }
      responses:
        "201": { description: Monitor created }
  /monitors/{id}:
    delete:
      summary: Delete monitor
      tags: [Monitor]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }

  # ── RAG ───────────────────────────────────────────────
  /rag/index:
    post:
      summary: Index a document
      tags: [RAG]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
                content: { type: string }
      responses:
        "200": { description: Indexed }
  /rag/search:
    get:
      summary: Semantic search
      tags: [RAG]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Search results

  # ── Task Engine ───────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...